requirement.txt.txt
streamlit==1.34.0
opencv-python==4.8.1.78
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
keras==2.15.0
Pillow==10.3.0
matplotlib==3.7.5
